Search: Fixes search limits and adds a page parameter
This adds a page parameter to search api without adding
any major breaking change.

It does at an api validation error when trying to use
a limit beyond 5000. This is a breaking change. We could
remove this and have it only in the docs and describe that this
is a limit that grafana will apply silently.

Fixes #16049

+++ pkg/services/dashboards/folder_service.go
@@ -9,7 +9,7 @@ import (
 
 // FolderService service for operating on folders
 type FolderService interface {
-	GetFolders(limit int) ([]*models.Folder, error)
+	GetFolders(limit int64) ([]*models.Folder, error)
 	GetFolderByID(id int64) (*models.Folder, error)
 	GetFolderByUID(uid string) (*models.Folder, error)
 	CreateFolder(cmd *models.CreateFolderCommand) error
@@ -25,7 +25,7 @@ var NewFolderService = func(orgId int64, user *models.SignedInUser) FolderServic
 	}
 }
 
-func (dr *dashboardServiceImpl) GetFolders(limit int) ([]*models.Folder, error) {
+func (dr *dashboardServiceImpl) GetFolders(limit int64) ([]*models.Folder, error) {
 	if limit == 0 {
 		limit = 1000
 	}
